refactor(#3278): move closure phase helpers to sibling closures/arrow-phases.ts (LOC-budget)
Slice 1 revised to the established WAVE B pattern (#3271/#3275/#3276): the two
extracted phases (planClosureCaptures capture analysis, mintClosureStructTypes
capture-struct type minting) live in a NEW sibling module
src/codegen/closures/arrow-phases.ts instead of inline, so the god-file SHRINKS
(#3102 LOC-budget gate: closures.ts 3472 -> 3129, was +70 inline). Exports four
shared private helpers (arrowOwnLocals, collectOverBody,
closureProvablyAfterLetDecl, buildCaptureFieldDef) for the sibling; the
resulting module cycle is safe (function-body-only usage).

Emitted-Wasm byte-identity re-proved IDENTICAL 39/39 (gc/standalone/wasi) vs
origin/main. tsc 0. tests/issue-3278.test.ts 10/10.

✓
